Move AzureAD component into its own file
