+++ src/index.tsx
@@ -23,113 +23,10 @@
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
 
-import * as React from 'react';
-import { Store } from 'redux';
-
-import { AAD_LOGIN_SUCCESS, AAD_LOGOUT_SUCCESS, loginSuccessful, logoutSuccessful } from './actions';
-import { IAuthProvider, IAuthProviderFactory, IUserInfo, LoginType, UserInfoCallback} from './Interfaces';
+import { AAD_LOGIN_SUCCESS, AAD_LOGOUT_SUCCESS } from './actions';
+import { AuthenticationState, AzureAD } from './AzureAD'
+import { LoginType } from './Interfaces';
 import { MsalAuthProviderFactory } from './MsalAuthProviderFactory';
-
-enum AuthenticationState {
-  Unauthenticated,
-  Authenticating,
-  Authenticated,
-}
-
-type UnauthenticatedFunction = (login: LoginFunction) => JSX.Element;
-type AuthenticatedFunction = (logout: LogoutFunction) => JSX.Element;
-type LoginFunction = () => void;
-type LogoutFunction = () => void;
-
-interface IProps {
-  provider : IAuthProviderFactory,
-  unauthenticatedFunction: UnauthenticatedFunction,
-  authenticatedFunction: AuthenticatedFunction,
-  userInfoCallback: UserInfoCallback,
-  reduxStore?: Store
-}
-
-interface IState {
-  authenticationState: AuthenticationState,
-}
-
-class AzureAD extends React.Component<IProps, IState> {
-  private authProvider : IAuthProvider;
-
-  constructor(props: IProps) {
-    super(props);
-
-    const authenticationState = AuthenticationState.Unauthenticated;
-
-    this.authProvider = this.props.provider.getAuthProvider();
-    this.authProvider.userInfoChangedCallback = this.updateState;
-
-    this.state = { authenticationState };
-  }
-
-  public render() {
-    switch (this.state.authenticationState) {
-      case AuthenticationState.Authenticated:
-        return this.props.authenticatedFunction(this.logout);
-      case AuthenticationState.Authenticating:
-        // TODO: Add loading callback, componentDidMount will acquire tokens and then re-render
-        return null;
-      case AuthenticationState.Unauthenticated:
-      default:
-        return this.props.unauthenticatedFunction(this.login);
-    }
-  }
-
-  public componentDidMount() {
-    this.sendUserInfo();
-  }
-
-  public sendUserInfo = (): void => {
-    const user : IUserInfo = this.authProvider.getUserInfo();
-    if (user) {
-      this.updateState(user);
-    }
-  }
-
-  public resetUserInfo = () => {
-    if (this.props.reduxStore) {
-      this.props.reduxStore.dispatch(logoutSuccessful());
-    }
-
-    this.setState({
-      authenticationState: AuthenticationState.Unauthenticated,
-    });
-  }
-
-  public updateState = (user: IUserInfo) => {
-    this.props.userInfoCallback(user);
-
-    this.dispatchToProvidedReduxStore(user);
-    this.setState({
-      authenticationState: AuthenticationState.Authenticated
-    })
-  }
-
-  private login = () => {
-    this.authProvider.login();
-    this.sendUserInfo();
-  };
-
-  private logout = () => {
-    if (this.state.authenticationState !== AuthenticationState.Authenticated) {
-      return;
-    }
-
-    this.resetUserInfo();
-    this.authProvider.logout();
-  };
-
-  private dispatchToProvidedReduxStore(data: IUserInfo) {
-    if (this.props.reduxStore) {
-      this.props.reduxStore.dispatch(loginSuccessful(data))
-    }
-  }
-}
 
 export { AzureAD, AAD_LOGIN_SUCCESS, AAD_LOGOUT_SUCCESS, AuthenticationState, LoginType, MsalAuthProviderFactory };
 export default AzureAD;
